# gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.DeleteProductErrorResponse:
    properties:
      error:
        example: <error_message>
        type: string
      message:
        example: Failed to Delete Product!
        type: string
    type: object
  models.DeleteProductSuccessResponse:
    properties:
      message:
        example: Product Deleted Successfully!
        type: string
    type: object
...
  models.ListProductsByVendorIDErrorResponse:
    properties:
      error:
        example: <error_message>
        type: string
      message:
        example: Failed to List Products!
        type: string
    type: object
  models.ListProductsByVendorIDSuccessResponse:
    properties:
      message:
        example: Products Listed Successfully!
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      total_count:
        example: 100
        type: integer
    type: object
...
  models.Product:
    properties:
      brand_id:
        example: 67adca0b4e864cfffb002299
        type: string
      category:
        example: Product Category
        type: string
      created_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      description:
        example: Product Description
        type: string
      id:
        example: 67adca0b4e864cfffb002299
        type: string
      inventory:
        $ref: '#/definitions/models.ProductInventory'
      name:
        example: Product Name
        type: string
      net_weight:
        example: 100
        type: number
      price:
        example: 100
        type: number
      updated_at:
        example: "2021-01-01T00:00:00Z"
        type: string
    type: object
  models.ProductCreateErrorResponse:
    properties:
      error:
        example: <error_message>
        type: string
      message:
        example: Failed to Create Product!
        type: string
    type: object
  models.ProductCreatePayload:
    properties:
      brand_id:
        example: 67adca0b4e864cfffb002299
        type: string
      category:
        example: Product Category
        type: string
      description:
        example: Product Description
        type: string
      name:
        example: Product Name
        type: string
      net_weight:
        example: 100
        type: number
      price:
        example: 100
        type: number
    required:
    - brand_id
    - description
    - name
    - net_weight
    - price
    type: object
  models.ProductCreateSuccessResponse:
    properties:
      message:
        example: Product Created Successfully!
        type: string
      product:
        $ref: '#/definitions/models.Product'
    type: object
  models.ProductInventory:
    properties:
      available_quantity:
        example: 100
        type: integer
      created_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      threshold_quantity:
        example: 10
        type: integer
      updated_at:
        example: "2021-01-01T00:00:00Z"
        type: string
    type: object
  models.ProductUpdateErrorResponse:
    properties:
      error:
        example: <error_message>
        type: string
      message:
        example: Failed to Update Product!
        type: string
    type: object
  models.ProductUpdatePayload:
    properties:
      brand_id:
        example: 67adca0b4e864cfffb002299
        type: string
      category:
        example: Product Category
        type: string
      description:
        example: Product Description
        type: string
      name:
        example: Product Name
        type: string
      net_weight:
        example: 100
        type: number
      price:
        example: 100
        type: number
    required:
    - brand_id
    - description
    - name
    - net_weight
    - price
    type: object
  models.ProductUpdateSuccessResponse:
    properties:
      message:
        example: Product Updated Successfully!
        type: string
      product:
        $ref: '#/definitions/models.Product'
    type: object
...
  models.UnauthorizedErrorResponse:
    properties:
      error:
        example: <error_message>
        type: string
      message:
        example: Unauthorized!
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: rohit.vilas.ingole@gmail.com
    name: Rohit Vilas Ingole
  description: This is the API for Gopher Social, a social media platform for Gophers.
  license:
    name: MIT License
    url: https://github.com/DataRohit/Order-Management-System/blob/master/license
  title: Gopher Social API
  version: "1.0"
paths:
...
  /product:
    get:
      consumes:
      - application/json
      description: List products for a vendor on the Platform with pagination
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Products Listed Successfully!
          schema:
            $ref: '#/definitions/models.ListProductsByVendorIDSuccessResponse'
        "400":
          description: Failed to List Products!
          schema:
            $ref: '#/definitions/models.ListProductsByVendorIDErrorResponse'
        "401":
          description: Unauthorized!
          schema:
            $ref: '#/definitions/models.UnauthorizedErrorResponse'
        "500":
          description: Internal Server Error!
          schema:
            $ref: '#/definitions/models.ListProductsByVendorIDErrorResponse'
      security:
      - BearerAuth: []
      summary: List products by vendor ID
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Create a new product on the Platform
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product Create Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.ProductCreatePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Product Created Successfully!
          schema:
            $ref: '#/definitions/models.ProductCreateSuccessResponse'
        "400":
          description: Failed to Create Product!
          schema:
            $ref: '#/definitions/models.ProductCreateErrorResponse'
        "401":
          description: Unauthorized!
          schema:
            $ref: '#/definitions/models.UnauthorizedErrorResponse'
        "500":
          description: Internal Server Error!
          schema:
            $ref: '#/definitions/models.ProductCreateErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Product
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product on the Platform
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product Deleted Successfully!
          schema:
            $ref: '#/definitions/models.DeleteProductSuccessResponse'
        "400":
          description: Failed to Delete Product!
          schema:
            $ref: '#/definitions/models.DeleteProductErrorResponse'
        "401":
          description: Unauthorized!
          schema:
            $ref: '#/definitions/models.UnauthorizedErrorResponse'
        "500":
          description: Internal Server Error!
          schema:
            $ref: '#/definitions/models.DeleteProductErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Update a product on the Platform
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product Update Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.ProductUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: Product Updated Successfully!
          schema:
            $ref: '#/definitions/models.ProductUpdateSuccessResponse'
        "400":
          description: Failed to Update Product!
          schema:
            $ref: '#/definitions/models.ProductUpdateErrorResponse'
        "401":
          description: Unauthorized!
          schema:
            $ref: '#/definitions/models.UnauthorizedErrorResponse'
        "500":
          description: Internal Server Error!
          schema:
            $ref: '#/definitions/models.ProductUpdateErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - Product
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"